Moved `host/syscall_types.rs` to `host/syscall/types.rs`
