+++ src/main/host/descriptor/epoll/mod.rs
@@ -14,7 +14,7 @@ use crate::host::descriptor::{
 };
 use crate::host::memory_manager::MemoryManager;
 use crate::host::syscall::io::IoVec;
-use crate::host::syscall_types::SyscallError;
+use crate::host::syscall::types::SyscallError;
 use crate::utility::callback_queue::CallbackQueue;
 use crate::utility::{HostTreePointer, ObjectCounter};
