+++ src/main/host/descriptor/socket/inet/mod.rs
@@ -18,7 +18,7 @@ use crate::host::memory_manager::MemoryManager;
 use crate::host::network::interface::FifoPacketPriority;
 use crate::host::network::namespace::{AssociationHandle, NetworkNamespace};
 use crate::host::syscall::io::IoVec;
-use crate::host::syscall_types::SyscallError;
+use crate::host::syscall::types::SyscallError;
 use crate::network::packet::PacketRc;
 use crate::utility::callback_queue::CallbackQueue;
 use crate::utility::sockaddr::SockaddrStorage;
