+++ src/main/host/descriptor/socket/unix.rs
@@ -22,7 +22,7 @@ use crate::host::descriptor::{
 use crate::host::memory_manager::MemoryManager;
 use crate::host::network::namespace::NetworkNamespace;
 use crate::host::syscall::io::{IoVec, IoVecReader, IoVecWriter};
-use crate::host::syscall_types::SyscallError;
+use crate::host::syscall::types::SyscallError;
 use crate::utility::callback_queue::CallbackQueue;
 use crate::utility::sockaddr::{SockaddrStorage, SockaddrUnix};
 use crate::utility::HostTreePointer;
