+++ src/main/host/descriptor/timerfd.rs
@@ -17,7 +17,7 @@ use crate::host::descriptor::{
 use crate::host::host::Host;
 use crate::host::memory_manager::MemoryManager;
 use crate::host::syscall::io::{IoVec, IoVecWriter};
-use crate::host::syscall_types::{SyscallError, SyscallResult};
+use crate::host::syscall::types::{SyscallError, SyscallResult};
 use crate::host::timer::Timer;
 use crate::utility::callback_queue::CallbackQueue;
 use crate::utility::HostTreePointer;
