+++ src/main/host/managed_thread.rs
@@ -25,7 +25,7 @@ use crate::core::scheduler;
 use crate::core::worker::{Worker, WORKER_SHARED};
 use crate::cshadow;
 use crate::host::syscall::handler::SyscallHandler;
-use crate::host::syscall_types::{ForeignArrayPtr, SyscallReturn};
+use crate::host::syscall::types::{ForeignArrayPtr, SyscallReturn};
 use crate::utility::{inject_preloads, syscall, verify_plugin_path, VerifyPluginPathError};
 
 /// The ManagedThread's state after having been allowed to execute some code.
