+++ src/main/host/mod.rs
@@ -10,6 +10,5 @@ pub mod process;
 pub mod status_listener;
 pub mod syscall;
 pub mod syscall_condition;
-pub mod syscall_types;
 pub mod thread;
 pub mod timer;
