+++ src/main/host/syscall/handler/fcntl.rs
@@ -7,7 +7,7 @@ use syscall_logger::log_syscall;
 use crate::cshadow;
 use crate::host::descriptor::{CompatFile, File, FileStatus};
 use crate::host::syscall::handler::{SyscallContext, SyscallHandler};
-use crate::host::syscall_types::SyscallResult;
+use crate::host::syscall::types::SyscallResult;
 
 impl SyscallHandler {
     #[log_syscall(/* rv */ std::ffi::c_int, /* fd */ std::ffi::c_int, /* cmd */ std::ffi::c_int)]
