+++ src/main/host/syscall/handler/ioctl.rs
@@ -8,7 +8,7 @@ use syscall_logger::log_syscall;
 use crate::cshadow as c;
 use crate::host::descriptor::{CompatFile, FileStatus};
 use crate::host::syscall::handler::{SyscallContext, SyscallHandler};
-use crate::host::syscall_types::SyscallResult;
+use crate::host::syscall::types::SyscallResult;
 
 impl SyscallHandler {
     #[log_syscall(/* rv */ std::ffi::c_int, /* fd */ std::ffi::c_int, /* request */ std::ffi::c_ulong)]
