+++ src/main/host/syscall/handler/poll.rs
@@ -3,7 +3,7 @@ use syscall_logger::log_syscall;
 
 use crate::cshadow as c;
 use crate::host::syscall::handler::{SyscallContext, SyscallHandler};
-use crate::host::syscall_types::SyscallError;
+use crate::host::syscall::types::SyscallError;
 
 impl SyscallHandler {
     #[log_syscall(/* rv */ std::ffi::c_int, /* ufds */ *const std::ffi::c_void,
