+++ src/main/host/syscall/io.rs
@@ -6,7 +6,7 @@ use linux_api::errno::Errno;
 use shadow_shim_helper_rs::syscall_types::ForeignPtr;
 
 use crate::host::memory_manager::MemoryManager;
-use crate::host::syscall_types::ForeignArrayPtr;
+use crate::host::syscall::types::ForeignArrayPtr;
 use crate::utility::sockaddr::SockaddrStorage;
 
 /// Writes the socket address into a buffer at `plugin_addr` with length `plugin_addr_len`, and
